tcmode: don't early expand PATH in its adjustments
To avoid issues when a user installed a Sourcery toolchain to the default path
and let it modify their PATH, we iterate over PATH and remove the problematic
elements. Unfortunately, in doing so, we inadvertently forced an early
expansion of PATH due to getting it in its expanded form, then setting PATH
based on that form. Instead, operate against the unexpanded PATH.

Signed-off-by: Christopher Larson <chris_larson@mentor.com>