README.md: add description of how the layer works
This information should be informative, covering exactly what the layer.conf and tcmode file are doing. In addition, it should address concerns about configuration migration from old versions of oe-core, without explicitly instructing the user in steps they may not need to perform.

This also adds instructions on the use of `external-sourcery-rebuild-libc`.