Separate out external-sourcery-toolchain components
This patch breaks up the external-sourcery-toolchain recipe into
parts, making the swap between prebuilt and recompiled eglibc easier
to see, and providing for the creation of an SDK (using a cross-canadian
recipe). It also adds optional support for mklibs and prelink, which
may be provided by some toolchains but not others.

The reason the base package is renamed is in part that there's also
an "external-sourcery-toolchain" recipe in oe-core, and in our build
system, this layer may be "under" oe-core, so we were picking up the
wrong one. The naming is negotiable.

Furthermore, fix multilib support: Instead of changing TARGET_PREFIX
to CSL_TARGET_SYS, we make the symlinks point from TARGET_PREFIX-util
to CSL_TARGET_SYS-util, and do this for both sets of multilibs. This
way, all the many things using ${TARGET_PREFIX} to distinguish multilibs
will work, e.g.:
     PREFERRED_PROVIDER_virtual/${TARGET_PREFIX}-gcc-initial

Since some toolchains have mklibs or prelink, and others don't, those
are installed conditionally based on PREFERRED_PROVIDER_prelink-native and
PREFERRED_PROVIDER_mklibs-native. Note that they go in the -cross recipe;
this is because, perhaps surprisingly, the Mentor toolchains contain a
slightly different prelink-rtld for each architecture.

The symlinking functionality is updated a bit, to have optional
symlinks (used for mklibs/prelink) and symlinks which will cause the
build to fail if they can't be made for a reason other than already
existing. It now searches three directories, and it's easier to modify
the logic.

There are probably loose ends to tie up. This is done as a single
large patch because many of the changes are interdependent.

Signed-off-by: Peter Seebach <peter.seebach@windriver.com>