glibc-external: library paths fix
the library paths in libpthread.so and libc.so points to system
libraries, which would not work in the ade. In the ade the relative
paths  are added so that correct libraries are linked. The fix is taken
from birch external-sourcery-toolchain recipe

recipe file at: https://github.com/MentorEmbedded/meta-
sourcery/blob/release/2014.12-async1/recipes/meta/external-sourcery-
toolchain.bb

Signed-off-by: adnan-ali1 <m_adnanali_1@hotmail.com>