glibc-external: add multilib link to -dev package
This link is required by linker to link libraries and crt* files in our
ADE/SDK. The 'rm -f' command is changed to 'rm -rf' because
external-toolchain.bbclass copies all the stuff required by PACKAGES to image
folder (in do_install stage). In doing so it copies the architecture folder
(as it is required by FILES_${PN}-dev ) from sourcery folder to image folder
in work directory. Afterwords when we try to remove the directory with 'rm -f'
it prompts an error as directories cannot be deleted with this command.

Signed-off-by: adnan-ali1 <adnan_ali@mentor.com>
Signed-off-by: Christopher Larson <chris_larson@mentor.com>