adds architecture link in ADE Target Sysroot
This link is required by linker to link libraries and crt* files. The
'rm -f' command is changed to 'rm -rf' because in cedar release
external-toolchain.bbclass copies all the stuff required by PACKAGES to
image folder (in do_install stage). In doing so it copies the
architecture folder (as it is required by FILES_glibc-multilib-link )
from sorcery folder to image folder in work directory. Afterwords when
we try to remove the directory with 'rm -f' it prompts an error as
directories cannot be deleted with this command.

Signed-off-by: adnan-ali1 <adnan_ali@mentor.com>