sourcery-tuning: remove -mcpu flag for t4240rdb-64b
Remove -mcpu=e6500 flag from TUNE_CCARGS for t4240rdb-64b
to match the toolchain options, in ADE's environment script,
with the CS3 data used by the IDE(Sourcery CodeBench).

IDE reads in the individual ABI and other toolchain options
from ADE's environment-setup script and attempts to map each
value to its appropriate OptionValue in the target data
plugin.xml. Many of the the values in the plugin.xml
(particularly the ABI values) originate from the CS3
information.

When the IDE finds a value in environment-setup script which
cannot be mapped to a value in plugin.xml, e.g -mcpu=e6500 in
this case, it simply appends it to the misc option field and
this behavior is known to be imperfect.

JIRA: SB-6290

Signed-off-by: Abdur Rehman <abdur_rehman@mentor.com>