Revert "Drop now-unnecessary <pkg>-dbg bits"
This reverts commit 91b303278497e23fb4644aabb0b70e541f71c7c0.

We need these explicitly listed in order to ensure we don't just package the
files, but extract them from the sysroot. The best approach would be to
identify elf binaries the way package.bbclass does, and extract the
debuglink path to determine the files to include, but that can be a future
enhancement.

JIRA: SB-8524
Signed-off-by: Christopher Larson <chris_larson@mentor.com>