glibc-sourcery: Add hash to glibc PKGVERSION
glibc includes some binaries such as ldd, that support a --version flag.
glibc can be configured with --with-pkgversion=<something> to display
some additional information for the --version flag.

We can calculate a hash value for the glibc sources and use it as
an argument for --with-pkgversion to sign our builds. That will allow us
to identify glibc binaries via the output of 'ldd --version'.

Signed-off-by: Mikhail Durnev <mikhail_durnev@mentor.com>