tcmode-external-sourcery: set SDKPATHTOOLCHAIN
This places the repackaged external toolchain in ${SDKPATH}/toolchains, with the
same structure as is used in the CodeBench toolchains directory. This is for
consistency with Yocto, which ensures that specifics are isolated in an SDK and
multiple SDKs can be installed to the same path without conflicts.

JIRA: SB-19962

Signed-off-by: Christopher Larson <chris_larson@mentor.com>