sdk-relocate-toolchain.bbclass: add a preceding newline to be safe
Rather than relying on toolchain-shar-relocate.sh having a trailing
newline to separate the content, ensure we have one here for safety.

With the current behavior, if toolchain-shar-relocate.sh lacks a trailing
newline at the end of the file, this will be directly concatenated to that, so
the first line of this function and the last of that file will be merged to a
single line, resulting in failures.

JIRA: SB-21318

Signed-off-by: Christopher Larson <chris_larson@mentor.com>